requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
google-generativeai==0.8.6
google-genai==1.67.0
google-api-python-client==2.164.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
cryptography==44.0.0
requests==2.32.3
